Remove container for SearchBox component; fix hydrate import error
